HSEARCH-1560 Moving batch indexing progress monitor impl to orm module.
