+++ orm/src/main/java/org/hibernate/search/batchindexing/impl/SimpleIndexingProgressMonitor.java
@@ -21,7 +21,7 @@
  * 51 Franklin Street, Fifth Floor
  * Boston, MA  02110-1301  USA
  */
-package org.hibernate.search.impl;
+package org.hibernate.search.batchindexing.impl;
 
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicLong;
