+++ orm/src/test/java/org/hibernate/search/test/errorhandling/ErrorHandlingDuringDocumentCreationTest.java
@@ -34,7 +34,7 @@ import org.hibernate.search.batchindexing.MassIndexerProgressMonitor;
 import org.hibernate.search.engine.spi.SearchFactoryImplementor;
 import org.hibernate.search.exception.ErrorHandler;
 import org.hibernate.search.test.SearchTestCaseJUnit4;
-import org.hibernate.search.testsupport.progessmonitor.AssertingMassIndexerProgressMonitor;
+import org.hibernate.search.test.util.progessmonitor.AssertingMassIndexerProgressMonitor;
 import org.hibernate.search.testsupport.TestForIssue;
 import org.jboss.byteman.contrib.bmunit.BMRule;
 import org.jboss.byteman.contrib.bmunit.BMUnitRunner;
